backend/requirements.txt
python-telegram-bot~=21.0
python-dotenv
fastapi
uvicorn[standard]
python-jose[cryptography]
